+++ apps/admin-x-settings/src/components/settings/membership/member-emails/use-welcome-email-preview.ts
@@ -1,10 +1,7 @@
 import {JSONError} from '@tryghost/admin-x-framework/errors';
 import {useCallback, useRef, useState} from 'react';
 
-export type WelcomeEmailDraft = {
-    subject: string;
-    lexical: string;
-};
+import {type WelcomeEmailDraft, getWelcomeEmailValidationErrors} from './welcome-email-validation';
 
 export type WelcomeEmailPreview = {
     html: string;
@@ -42,45 +39,6 @@ const preparePreviewHtml = (html: string) => {
     });
 
     return `<!doctype html>${parsed.documentElement.outerHTML}`;
-};
-
-const isEmptyLexical = (lexical: string | null | undefined): boolean => {
-    if (!lexical) {
-        return true;
-    }
-
-    try {
-        const parsed = JSON.parse(lexical);
-        const children = parsed?.root?.children;
-
-        // Empty if no children or only an empty paragraph
-        if (!children || children.length === 0) {
-            return true;
-        }
-        if (children.length === 1 &&
-            children[0].type === 'paragraph' &&
-            (!children[0].children || children[0].children.length === 0)) {
-            return true;
-        }
-
-        return false;
-    } catch {
-        return true;
-    }
-};
-
-export const getWelcomeEmailValidationErrors = (state: WelcomeEmailDraft): Record<string, string> => {
-    const newErrors: Record<string, string> = {};
-
-    if (!state.subject?.trim()) {
-        newErrors.subject = 'A subject is required';
-    }
-
-    if (isEmptyLexical(state.lexical)) {
-        newErrors.lexical = 'Email content is required';
-    }
-
-    return newErrors;
 };
 
 const getPreviewErrorMessage = (error: unknown) => {
